rename files, edit comments, unset workspace store on route change
