+++ ui/src/app/pages/access/access-renewal.spec.tsx
@@ -9,7 +9,7 @@ import SpyInstance = jest.SpyInstance;
 import {InstitutionApiStub} from 'testing/stubs/institution-api-stub';
 import {ProfileApiStub} from 'testing/stubs/profile-api-stub';
 import {ProfileStubVariables} from 'testing/stubs/profile-api-stub';
-import {AccessRenewal} from 'app/pages/access/access-renewal-page';
+import {AccessRenewal} from 'app/pages/access/access-renewal';
 import {findNodesByExactText, findNodesContainingText, waitOneTickAndUpdate} from 'testing/react-test-helpers';
 
 const EXPIRY_DAYS = 365
