+++ ui/src/app/pages/profile/profile.spec.tsx
@@ -2,7 +2,7 @@ import {mount, ReactWrapper} from 'enzyme';
 import * as React from 'react';
 
 import {TextInput} from 'app/components/inputs';
-import {ProfileComponent} from 'app/pages/profile/profile-page';
+import {ProfileComponent} from 'app/pages/profile/profile';
 import {registerApiClient} from 'app/services/swagger-fetch-clients';
 import {profileStore, serverConfigStore} from 'app/utils/stores';
 import {InstitutionApi, ProfileApi} from 'generated/fetch';
